Move src to root
