+++ platformio.ini
@@ -16,7 +16,6 @@ include_dir = .
 
 [env]
 build_flags = 
-	-I./src
 	-Wall
 	-std=c++17
 lib_deps =
